requirements.txt
discord.py==1.2.3
GitPython==2.1.13
python-dotenv==0.10.3
aiosqlite==0.10.0
psutil==5.6.3
lyricsgenius==1.7.0
jishaku==1.16.1.147
PyYAML==5.1.2
psycopg2==2.8.3
asyncpg==0.18.3
